dom/url/tests/test_urlutils_stringify.html


<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=959190
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 959190</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=959190">Mozilla Bug 959190</a>
<p id="display"></p>
<div id="content" style="display: none">
  <iframe name="x" id="x"></iframe>
  <iframe name="y" id="y"></iframe>
</div>
<pre id="test">
</pre>
  <a id="link" href="http://www.example.com:8080">foobar</a>
  <area id="area" href="http://www.example.com:8080" />
  <script type="application/javascript">

  var url = new URL("http://www.example.com:8080");
  is(url + "", "http://www.example.com:8080/", "URL stringify");

  var link = document.getElementById("link");
  is(link + "", "http://www.example.com:8080/", "Anchor stringify");

  var area = document.getElementById("area");
  is(area + "", "http://www.example.com:8080/", "Area stringify");

  is((location + "").indexOf("http://mochi.test:8888/tests/dom/url/tests/test_urlutils_stringify.html"), 0, "Location stringify");
  </script>
</body>
</html>
